Fix building of spiffs_service
We encountered serveral issues, first the extern "C" marks were not set
correctly which lead to some confusion with the calling of C functions
from XC.
Secondly (and more seriously) the checker at the end of the build is not
able the guess the stack size of C functions correctly. For this the
pragmas `stackfunction` and `stackcalls` are introduced to give enough
hints about the used stack size. With this patch it is actually possible
to run the spiffs service on the same tile as the flash service.

FIXME Needs fine tuning!